+++ API/DTOs/Wallet/ChangeWalletCoinDto.cs
@@ -8,7 +8,7 @@ namespace API.DTOs;
 public class ChangeWalletCoinDto
 {
     public int appUserID { get; set; }
-    public float BalanceBuy { get; set; }
+    public double BalanceBuy { get; set; }
     public int coinIDFrom { get; set; }
     public int coinIDTo { get; set; }
 }
