+++ API/Model/Wallet/Coin.cs
@@ -9,7 +9,7 @@ public class Coin{
 
     public int Id { get; set; }
     public string Name { get; set; }
-    public float ConvertionToEuro { get; set; }
+    public double ConvertionToEuro { get; set; }
     public ICollection<WalletCoin> WalletCoin { get; set; }    
 
 }
